se creo funcionamiento sin ajax del servidor nodejs

+++ node.php
@@ -1,20 +1,17 @@
 <?php
 
-class Node{
+define("UNDEFINE", 0);
+define("ERROR", 1);
+define("START", 2);
+define("RUNNING", 3);
+define("NORUNNING", 4);
+define("STOP", 5);
 
-  define("UNDEFINE", -1);
-  define("ERROR", 0);
-  define("START", 1);
-  define("RUNNING", 2);
-  define("NORUNNING", 3);
-  define("STOP", 4);
+class Node{
 
   private $NODE_DIR, $NODE_PORT, $NODE_PID_DIR, $ADMIN_PASS;
 
   public function Node($NODE_DIR, $NODE_PORT, $NODE_PID_DIR, $ADMIN_PASS) {
-    error_reporting(E_ALL); //Define tipos de errores a ejecutar.
-    set_time_limit(120); //Tiempo maximo de ejecucion.
-
     $this->NODE_DIR     = $NODE_DIR; //Carpeta de Nodejs.
     $this->NODE_PORT    = $NODE_PORT; //Puerto del servidor Nodejs.
     $this->NODE_PID_DIR = $NODE_PID_DIR; //Carpeta donde se almacena direccion del proceso.
@@ -44,13 +41,15 @@ class Node{
 
     if(!file_exists($this->NODE_DIR)) return $this->report(UNDEFINE);
 
-  	$node_pid = intval(file_get_contents("$this->NODE_PID_DIR/nodepid"));
+  	$node_pid = @intval(file_get_contents("$this->NODE_PID_DIR/nodepid"));
 
   	if($node_pid > 0) return $this->report(RUNNING);
 
+    if(!file_exists($FILE)) return $this->report(ERROR);
+
   	$file = escapeshellarg($FILE);
 
-  	$node_pid = exec("PORT=$this->NODE_PORT $this->NODE_DIR /bin/node $file >$this->NODE_PID_DIR/nodepid 2>&1 & echo $!");
+  	$node_pid = exec("PORT=$this->NODE_PORT $this->NODE_DIR/bin/node $file > /dev/null & echo $!");
 
     file_put_contents("$this->NODE_PID_DIR/nodepid", $node_pid, LOCK_EX);
   	sleep(1); //Espera que se levante Notejs.
@@ -65,16 +64,18 @@ class Node{
 
     if(!file_exists($this->NODE_DIR)) return $this->report(UNDEFINE);
 
-    $node_pid = intval(file_get_contents("$this->NODE_PID_DIR/nodepid"));
+    $node_pid = @intval(file_get_contents("$this->NODE_PID_DIR/nodepid"));
 
   	if($node_pid === 0) return $this->report(NORUNNING);
 
   	$ret = -1;
   	passthru("kill $node_pid", $ret);
-    file_put_contents("$this->NODE_PID_DIR/nodepid", '', LOCK_EX);
 
-    if($ret === 0) return $this->report(STOP);
-    else return $this->report(ERROR);
+
+    if($ret === 0){
+      file_put_contents("$this->NODE_PID_DIR/nodepid", '', LOCK_EX);
+      return $this->report(STOP);
+    }else return $this->report(ERROR);
   }
 
 }
